Move UserProfileLayout into new folder
